Added DefaultDiffAdapter that calculates diff on main thread
